requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
playwright==1.46.0
matplotlib==3.9.2
pyyaml==6.0.2
xgboost==2.1.1
sqlite3